requirement.txt
asgiref==3.3.1
Django==3.1.5
djangorestframework==3.12.2
install==1.3.4
Pygments==2.7.4
pytz==2020.5
sqlparse==0.4.1